% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_generate_simulationData_homogeneous_design.R
\name{gen_simData_homo}
\alias{gen_simData_homo}
\title{Simulated Dataset Generation for High-Dimensional Mediation Analysis}
\usage{
gen_simData_homo(
  n = 100,
  p_x = 5,
  rho = 0,
  p_m = 100,
  h = 0,
  source = FALSE,
  transferable = TRUE,
  seed = NULL
)
}
\arguments{
\item{n}{Integer. Number of observations (sample size). Default is 100.}

\item{p_x}{Integer. Number of covariates (confounders). Default is 5.}

\item{rho}{Numeric. Correlation coefficient (0–1) controlling correlation between mediators. Default is 0 (no correlation).}

\item{p_m}{Integer. Number of mediators. Default is 100.}

\item{h}{Integer. Degree of heterogeneity (for source data). Default is 0.}

\item{source}{Logical. If TRUE, generate source (external) dataset. Default is FALSE.}

\item{transferable}{Logical. If TRUE, generates a transferable source dataset sharing mediator-outcome structure with the target. Default is TRUE.}

\item{seed}{Integer. Optional random seed for reproducibility. Default is NULL.}
}
\value{
A list with the following components:
\itemize{
\item{\code{data}}: A data.frame of dimension \code{n * (2 + p_m + p_x)} containing outcome (Y), treatment (D), mediators (M1–Mp_m), and covariates (X1–Xp_x).
\item{\code{coef}}: A named list of true model coefficients (\code{alpha1, alpha2, beta1, beta2, beta3, beta4}, etc.).
}
}
\description{
Generates synthetic datasets mimicking high-dimensional mediation structures,
optionally incorporating transferable source data under varying covariate correlation
and heterogeneity levels. This function supports homogeneous settings for data
generation.
}
\details{
This function generates data according to a structural equation model (SEM):
\itemize{
\item Treatment D depends linearly on covariates X.
\item Mediators M depend on D and X, with residual correlation controlled by \code{rho}.
\item Outcome Y depends on D, M, and X.
\item If \code{source = TRUE}, then a source dataset is simulated with potentially transferable mechanisms.
}
}
\examples{
source_data <- gen_simData_homo(
  n = 100, p_x = 5, rho = 0, p_m = 100, h = 0,
  source = TRUE, transferable = TRUE, seed = 123
)
target_data <- gen_simData_homo(
  n = 100, p_x = 5, rho = 0, p_m = 100, h = 0,
  source = FALSE, transferable = TRUE, seed = 123
)

}
