% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_5.R
\name{summary.dblasso}
\alias{summary.dblasso}
\title{Summary of Debiased Lasso Inference}
\usage{
\method{summary}{dblasso}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"dblasso"}.}

\item{...}{Further arguments (currently not used).}
}
\value{
An object of class \code{"summary.dblasso"}.
}
\description{
Summary of Debiased Lasso Inference
}
