% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_4.R
\name{summary.lasso}
\alias{summary.lasso}
\title{Summary of Lasso Regression}
\usage{
\method{summary}{lasso}(object, ...)
}
\arguments{
\item{object}{A numeric vector of lasso coefficients with names.}

\item{...}{Further arguments (currently not used).}
}
\value{
An object of class \code{"summary.lasso"}.
}
\description{
Summary of Lasso Regression
}
