% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_7.R
\name{summary.source_detection}
\alias{summary.source_detection}
\title{Summary of Source Detection Results}
\usage{
\method{summary}{source_detection}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"source_detection"}.}

\item{...}{Further arguments (unused).}
}
\value{
An object of class \code{"summary.source_detection"}.
}
\description{
Summary of Source Detection Results
}
