% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_9.R
\name{summary.TransHDM}
\alias{summary.TransHDM}
\title{Summary of TransHDM Mediation Analysis}
\usage{
\method{summary}{TransHDM}(object, top = 10, digits = 4, ...)
}
\arguments{
\item{object}{An object of class \code{"TransHDM"}.}

\item{top}{Integer, maximum number of mediators to display.}

\item{digits}{Number of digits for rounding estimates.}

\item{...}{Further arguments (unused).}
}
\value{
An object of class \code{"summary.TransHDM"}.
}
\description{
Summary of TransHDM Mediation Analysis
}
